employee_vacation_service/requirements.txt
Django==4.0
psycopg2-binary==2.9.3
django-environ==0.9.0
djangorestframework==3.13.1
drf-yasg==1.21.4